requirements.txt
fastapi==0.128.0
joblib==1.5.2
numpy==2.3.5
pandas==2.3.3
pydantic==2.12.5
pyyaml==6.0.2
scikit-learn==1.7.2
uvicorn==0.40.0
